requirements.txt
streamlit>=1.32.0
langchain>=0.1.10
langchain-core>=0.1.28
langchain-community>=0.0.24
langchain-chroma>=0.1.0
chromadb>=0.4.22
dashscope>=1.14.0
python-dotenv>=1.0.1
requests>=2.31.0